Updated files with `git_helper` and trialling new .travis.yml layout

+++ .ci/travis_deploy_conda.sh
@@ -24,6 +24,7 @@ if [ $TRAVIS_PYTHON_VERSION == 3.6 ]; then
   conda build conda -c domdfcoding -c conda-forge --output-folder conda/dist --skip-existing
 
   for f in conda/dist/noarch/domdf_python_tools-*.tar.bz2; do
+    [ -e "$f" ] || continue
     echo "$f"
     conda install $f || exit 1
     if [ -z "$TRAVIS_TAG" ]; then
